1563 - Convert floats to 32

+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1634,7 +1634,7 @@ class ModelRateOfChangeSchemas:
         IndCQC.care_home: pl.String,
         IndCQC.number_of_beds: pl.Int32,
         IndCQC.primary_service_type: pl.String,
-        IndCQC.combined_ratio_and_filled_posts: pl.Float64,
+        IndCQC.combined_ratio_and_filled_posts: pl.Float32,
         IndCQC.care_home_status_count: pl.Int32,
     }
     expected_model_roc_trendline_schema = {
@@ -1643,9 +1643,9 @@ class ModelRateOfChangeSchemas:
         IndCQC.care_home: pl.String,
         IndCQC.number_of_beds: pl.Int32,
         IndCQC.primary_service_type: pl.String,
-        IndCQC.combined_ratio_and_filled_posts: pl.Float64,
+        IndCQC.combined_ratio_and_filled_posts: pl.Float32,
         IndCQC.care_home_status_count: pl.Int32,
-        IndCQC.ascwds_rate_of_change_trendline_model: pl.Float64,
+        IndCQC.ascwds_rate_of_change_trendline_model: pl.Float32,
     }
 
     calculate_rolling_sums_schema = {
@@ -1653,33 +1653,33 @@ class ModelRateOfChangeSchemas:
         IndCQC.primary_service_type: pl.String,
         IndCQC.number_of_beds_banded_roc: pl.Int32,
         IndCQC.cqc_location_import_date: pl.Date,
-        ROC_TempCol.current_period_cleaned: pl.Float64,
-        ROC_TempCol.previous_period_cleaned: pl.Float64,
+        ROC_TempCol.current_period_cleaned: pl.Float32,
+        ROC_TempCol.previous_period_cleaned: pl.Float32,
     }
     expected_calculate_rolling_sums_schema = {
         IndCQC.location_id: pl.String,
         IndCQC.primary_service_type: pl.String,
         IndCQC.number_of_beds_banded_roc: pl.Int32,
         IndCQC.cqc_location_import_date: pl.Date,
-        ROC_TempCol.rolling_current_sum: pl.Float64,
-        ROC_TempCol.rolling_previous_sum: pl.Float64,
+        ROC_TempCol.rolling_current_sum: pl.Float32,
+        ROC_TempCol.rolling_previous_sum: pl.Float32,
     }
 
     clean_non_residential_rate_of_change_schema = {
         IndCQC.location_id: pl.String,
         IndCQC.care_home: pl.String,
         IndCQC.cqc_location_import_date: pl.Date,
-        ROC_TempCol.previous_period_interpolated: pl.Float64,
-        ROC_TempCol.current_period_interpolated: pl.Float64,
+        ROC_TempCol.previous_period_interpolated: pl.Float32,
+        ROC_TempCol.current_period_interpolated: pl.Float32,
     }
     expected_clean_non_residential_rate_of_change_schema = {
         IndCQC.location_id: pl.String,
         IndCQC.care_home: pl.String,
         IndCQC.cqc_location_import_date: pl.Date,
-        ROC_TempCol.previous_period_interpolated: pl.Float64,
-        ROC_TempCol.current_period_interpolated: pl.Float64,
-        ROC_TempCol.previous_period_cleaned: pl.Float64,
-        ROC_TempCol.current_period_cleaned: pl.Float64,
+        ROC_TempCol.previous_period_interpolated: pl.Float32,
+        ROC_TempCol.current_period_interpolated: pl.Float32,
+        ROC_TempCol.previous_period_cleaned: pl.Float32,
+        ROC_TempCol.current_period_cleaned: pl.Float32,
     }
 
     calculate_trendline_schema = {
